Add dynamic import on zxcvbn
Change-type: patch
